Extract asset reading to AssetReader
